+++ merged_parallel/rayTrace_gpu.h
@@ -29,40 +29,40 @@ Point3 computeCameraPosition(const Point3& leftCorner, const Point3& rightCorner
 // Device-side Moller-Trumbore intersection
 inline __device__
 HitRecord mollerTrumbore(const Ray& ray, const Triangle& tri) {
-    const double EPSILON = 1e-8;
+    const float EPSILON = 1e-8f;
 
-    double e1x = tri.v2.x - tri.v1.x;
-    double e1y = tri.v2.y - tri.v1.y;
-    double e1z = tri.v2.z - tri.v1.z;
+    float e1x = tri.v2.x - tri.v1.x;
+    float e1y = tri.v2.y - tri.v1.y;
+    float e1z = tri.v2.z - tri.v1.z;
 
-    double e2x = tri.v3.x - tri.v1.x;
-    double e2y = tri.v3.y - tri.v1.y;
-    double e2z = tri.v3.z - tri.v1.z;
+    float e2x = tri.v3.x - tri.v1.x;
+    float e2y = tri.v3.y - tri.v1.y;
+    float e2z = tri.v3.z - tri.v1.z;
 
-    double pvecX = ray.direction.y * e2z - ray.direction.z * e2y;
-    double pvecY = ray.direction.z * e2x - ray.direction.x * e2z;
-    double pvecZ = ray.direction.x * e2y - ray.direction.y * e2x;
+    float pvecX = ray.direction.y * e2z - ray.direction.z * e2y;
+    float pvecY = ray.direction.z * e2x - ray.direction.x * e2z;
+    float pvecZ = ray.direction.x * e2y - ray.direction.y * e2x;
 
-    double det = e1x * pvecX + e1y * pvecY + e1z * pvecZ;
-    if (std::fabs(det) < EPSILON) return HitRecord();
+    float det = e1x * pvecX + e1y * pvecY + e1z * pvecZ;
+    if (fabsf(det) < EPSILON) return HitRecord();
 
-    double invDet = 1.0 / det;
+    float invDet = 1.0f / det;
 
-    double tvecX = ray.origin.x - tri.v1.x;
-    double tvecY = ray.origin.y - tri.v1.y;
-    double tvecZ = ray.origin.z - tri.v1.z;
+    float tvecX = ray.origin.x - tri.v1.x;
+    float tvecY = ray.origin.y - tri.v1.y;
+    float tvecZ = ray.origin.z - tri.v1.z;
 
-    double u = (tvecX * pvecX + tvecY * pvecY + tvecZ * pvecZ) * invDet;
+    float u = (tvecX * pvecX + tvecY * pvecY + tvecZ * pvecZ) * invDet;
     if (u < 0.0 || u > 1.0) return HitRecord();
 
-    double qvecX = tvecY * e1z - tvecZ * e1y;
-    double qvecY = tvecZ * e1x - tvecX * e1z;
-    double qvecZ = tvecX * e1y - tvecY * e1x;
+    float qvecX = tvecY * e1z - tvecZ * e1y;
+    float qvecY = tvecZ * e1x - tvecX * e1z;
+    float qvecZ = tvecX * e1y - tvecY * e1x;
 
-    double baryV = (ray.direction.x * qvecX + ray.direction.y * qvecY + ray.direction.z * qvecZ) * invDet;
+    float baryV = (ray.direction.x * qvecX + ray.direction.y * qvecY + ray.direction.z * qvecZ) * invDet;
     if (baryV < 0.0 || u + baryV > 1.0) return HitRecord();
 
-    double rayT = (e2x * qvecX + e2y * qvecY + e2z * qvecZ) * invDet;
+    float rayT = (e2x * qvecX + e2y * qvecY + e2z * qvecZ) * invDet;
     if (rayT < EPSILON) return HitRecord();
 
     Point3 intersection(
@@ -80,7 +80,7 @@ HitRecord findIntersectingTriangle(const Ray& ray,
                                    Triangle* sceneTriangles,
                                    int* cellStart,
                                    int boxNumber) {
-    double closestDistance = INFINITY;
+    float closestDistance = INFINITY;
     HitRecord closestHit;
     closestHit.hit = false;
     int start = cellStart[boxNumber];
